requirements.txt
alabaster==0.7.9
appnope==0.1.0
astroid==1.4.7
Babel==2.3.4
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
blinker==1.4
boto==2.45.0
bz2file==0.98
click==6.6
configparser==3.5.0
coverage==4.2
decorator==4.0.10
docutils==0.12
enum34==1.1.6
Flask==0.11.1
Flask-Babel==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.13.1
flipflop==1.0
gensim==0.13.4
guess-language==0.2
httplib2==0.9.2
imagesize==0.7.1
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.0
lazy-object-proxy==1.2.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.1.0
nltk==3.2.1
numpy==1.11.3
oauth==1.0.1
oauth2==1.9.0.post1
pandas==0.19.0
pathlib2==2.1.0
pbr==1.10.0
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.4
prompt-toolkit==1.0.3
psutil==4.3.1
ptyprocess==0.5.1
pyenchant==1.6.8
pyflakes==1.3.0
Pygments==2.1.3
pylint==1.5.4
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2016.7
pyzmq==16.0.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
rauth==0.7.2
requests==2.12.4
rope==0.9.4
scipy==0.18.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.3.5
snowballstemmer==1.2.1
Sphinx==1.4.8
spyder==3.0.1
SQLAlchemy==1.1.2
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
Whoosh==2.7.4
wrapt==1.10.8
WTForms==2.1
xlrd==1.0.0
